requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
google-auth==2.34.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-text-splitters==0.2.4
langsmith==0.1.112
multidict==6.0.5
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlparse==0.5.1
tenacity==8.5.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.9
